Move chunk-backed RandomAccessFile implementation
