+++ libvast/native-plugins/feather.cpp
@@ -25,154 +25,10 @@ namespace vast::plugins::feather {
 
 namespace {
 
-class chunk_random_access_file final : public arrow::io::RandomAccessFile {
-public:
-  explicit chunk_random_access_file(chunk_ptr chunk)
-    : chunk_(std::move(chunk)) {
-  }
-
-  // TODO: all methods, check position < size
-  arrow::Result<int64_t> GetSize() override {
-    return detail::narrow_cast<int64_t>(chunk_->size());
-  }
-
-  arrow::Result<int64_t>
-  ReadAt(int64_t position, int64_t nbytes, void* out) override {
-    const auto clamped_size
-      = std::min(chunk_->size() - detail::narrow_cast<size_t>(position),
-                 detail::narrow_cast<size_t>(nbytes));
-    const auto bytes = as_bytes(chunk_).subspan(position, clamped_size);
-    std::memcpy(out, bytes.data(), bytes.size());
-    return bytes.size();
-  }
-
-  arrow::Result<std::shared_ptr<arrow::Buffer>>
-  ReadAt(int64_t position, int64_t nbytes) override {
-    const auto clamped_size
-      = std::min(chunk_->size() - detail::narrow_cast<size_t>(position),
-                 detail::narrow_cast<size_t>(nbytes));
-    return as_arrow_buffer(chunk_->slice(position, clamped_size));
-  }
-
-  // /// EXPERIMENTAL: Read data asynchronously.
-  arrow::Future<std::shared_ptr<arrow::Buffer>>
-  ReadAsync(const arrow::io::IOContext&, int64_t position,
-            int64_t nbytes) override {
-    return arrow::Future<std::shared_ptr<arrow::Buffer>>::MakeFinished(
-      ReadAt(position, nbytes));
-  }
-
-  arrow::Status WillNeed(const std::vector<arrow::io::ReadRange>&) override {
-    return arrow::Status::OK();
-  }
-
-  arrow::Status Seek(int64_t position) override {
-    position_ = position;
-    return arrow::Status::OK();
-  }
-
-  arrow::Result<arrow::util::string_view> Peek(int64_t nbytes) override {
-    const auto clamped_size = std::min(chunk_->size() - position_,
-                                       detail::narrow_cast<size_t>(nbytes));
-    const auto bytes = as_bytes(chunk_).subspan(position_, clamped_size);
-    return arrow::util::string_view{reinterpret_cast<const char*>(bytes.data()),
-                                    bytes.size()};
-  }
-
-  /// \brief Return true if InputStream is capable of zero copy Buffer reads
-  ///
-  /// Zero copy reads imply the use of Buffer-returning Read() overloads.
-  bool supports_zero_copy() const override {
-    return true;
-  }
-
-  /// \brief Read and return stream metadata
-  ///
-  /// If the stream implementation doesn't support metadata, empty metadata
-  /// is returned.  Note that it is allowed to return a null pointer rather
-  /// than an allocated empty metadata.
-  arrow::Result<std::shared_ptr<const arrow::KeyValueMetadata>>
-  ReadMetadata() override {
-    return nullptr;
-  }
-
-  /// \brief Read stream metadata asynchronously
-  arrow::Future<std::shared_ptr<const arrow::KeyValueMetadata>>
-  ReadMetadataAsync(const arrow::io::IOContext&) override {
-    return arrow::Future<
-      std::shared_ptr<const arrow::KeyValueMetadata>>::MakeFinished(nullptr);
-  }
-
-  arrow::Status Close() override {
-    closed_ = true;
-    return arrow::Status::OK();
-  }
-
-  /// \brief Close the stream asynchronously
-  ///
-  /// By default, this will just submit the synchronous Close() to the
-  /// default I/O thread pool. Subclasses may implement this in a more
-  /// efficient manner.
-  arrow::Future<> CloseAsync() override {
-    // TODO: necessary?
-    return arrow::Future<>::MakeFinished(Close());
-  };
-
-  /// \brief Close the stream abruptly
-  ///
-  /// This method does not guarantee that any pending data is flushed.
-  /// It merely releases any underlying resource used by the stream for
-  /// its operation.
-  ///
-  /// After Abort() is called, closed() returns true and the stream is not
-  /// available for further operations.
-  arrow::Status Abort() override {
-    return Close();
-  }
-
-  /// \brief Return the position in this stream
-  arrow::Result<int64_t> Tell() const override {
-    return detail::narrow_cast<int64_t>(position_);
-  }
-
-  /// \brief Return whether the stream is closed
-  bool closed() const override {
-    return closed_;
-  }
-
-  arrow::Result<int64_t> Read(int64_t nbytes, void* out) override {
-    return ReadAt(detail::narrow_cast<int64_t>(position_), nbytes, out);
-  }
-
-  /// \brief Read data from current file position.
-  ///
-  /// Read at most `nbytes` from the current file position. Less bytes may
-  /// be read if EOF is reached. This method updates the current file position.
-  ///
-  /// In some cases (e.g. a memory-mapped file), this method may avoid a
-  /// memory copy.
-  arrow::Result<std::shared_ptr<arrow::Buffer>> Read(int64_t nbytes) override {
-    return ReadAt(detail::narrow_cast<int64_t>(position_), nbytes);
-  }
-
-  /// EXPERIMENTAL: The IOContext associated with this file.
-  ///
-  /// By default, this is the same as default_io_context(), but it may be
-  /// overriden by subclasses.
-  const arrow::io::IOContext& io_context() const override {
-    return arrow::io::default_io_context();
-  }
-
-private:
-  chunk_ptr chunk_;
-  size_t position_ = {};
-  bool closed_ = false;
-};
-
 class passive_feather_store final : public passive_store {
   [[nodiscard]] caf::error load(chunk_ptr chunk) override {
     std::shared_ptr<arrow::Table> table = {};
-    auto file = std::make_shared<chunk_random_access_file>(chunk);
+    auto file = as_arrow_file(std::move(chunk));
     auto reader = arrow::ipc::feather::Reader::Open(file).ValueOrDie();
     auto status = reader->Read(&table);
     VAST_ASSERT(status.ok());
